Update README.md
Instructions for use. exctract all files (the .exe and the config file) into the same directory and run the ,exe. select the folder where your code is and click start tracking. you can watch it work in the console and check the status in file status as well open the files directly for editing.